Reorganize sources and move to Meson
The sources now mimic the structure of '/usr/include/glib-2.0' to
provide maximum backward compatibility.

Drop posixvalac in favour of two step compilation:

 1. compile to C using valac with usual glib-2.0 and gobject-2.0 vapis
 2. compile executables with the glib-posix-2.0 and gobject-posix-2.0 stubs

In term of command line:

 1. valac --ccode --pkg=glib-posix-2.0 --pkg=gobject-posix-2.0 main.vala
 2. gcc $(pkg-config --cflags --libs glib-posix-2.0 gobject-posix-2.0) main.c

Split posixvala into two stubbed dependencies:

 - glib-posix-2.0 which reimplements glib-2.0
 - gobject-posix-2.0 which reimplements gobject-2.0

Install empty VAPIs for both 'glib-posix-2.0' and 'gobject-posix-2.0' so
that Meson does not complains as it sets '--pkg' flags.

Add '.deps' file even if it's implicit. This way, using '--nostdpkg'
with posixified VAPIs would simply include the definitions in
'glib-2.0.vapi' and 'gobject-2.0.vapi'.
