+++ include/glib-object.h
@@ -1,0 +1,8 @@
+#ifndef __GLIB_OBJECT_H__
+#define __GLIB_OBJECT_H__
+
+#include "gobject/gboxed.h"
+#include "gobject/gtype.h"
+#include "gobject/gobject.h"
+
+#endif /* __GLIB_OBJECT_H__ */
