+++ include/glib/gstring.h
