+++ include/gobject/gobject.h
@@ -1,0 +1,2 @@
+typedef void* GObject;
+typedef void* GObjectClass;
