+++ include/gobject/gtype.h
@@ -1,0 +1,10 @@
+#ifndef __G_TYPE__
+#define __G_TYPE__
+
+typedef int GType;
+
+typedef void* GTypeInterface;
+
+static inline void g_type_init () {}
+
+#endif
