+++ include/meson.build
@@ -1,0 +1,19 @@
+glib_posix_headers = [
+    'glib/gatomic.h',
+    'glib/gerror.h',
+    'glib/glist.h',
+    'glib/gstring.h',
+    'glib/gslice.h'
+]
+
+install_headers(glib_posix_headers, subdir: 'glib-posix-2.0/glib')
+
+gobject_posix_headers = [
+    'gobject/gboxed.h',
+    'gobject/gtype.h',
+    'gobject/gvaluecollector.h'
+]
+
+install_headers(gobject_posix_headers, subdir: 'glib-posix-2.0/gobject')
+
+install_headers(['glib.h', 'glib-object.h'], subdir: 'glib-posix-2.0')
