--- pkgconfig/posixvala.pc
@@ -1,7 +1,0 @@
-prefix=/usr/local
-includedir=${prefix}/include/
- 
-Name: posixvala
-Description: Posix profile for Vala
-Version: 0.2
-Cflags: -I${includedir}/posixvala
