+++ tests/posix.vala
@@ -1,0 +1,5 @@
+using Posix;
+
+int main () {
+	return Posix.EXIT_SUCCESS;
+}
