+++ vapi/glib-posix-2.0.deps
@@ -1,0 +1,1 @@
+glib-2.0
