+++ vapi/gobject-posix-2.0.deps
@@ -1,0 +1,1 @@
+gobject-2.0
