+++ vapi/gobject-posix-2.0.vapi
@@ -1,0 +1,3 @@
+namespace GLib {
+
+}
