+++ vapi/meson.build
@@ -1,0 +1,6 @@
+install_data([
+    'glib-posix-2.0.deps',
+    'glib-posix-2.0.vapi',
+    'gobject-posix-2.0.deps',
+    'gobject-posix-2.0.vapi'
+], install_dir: 'share/vala/vapi')
